Add php-mysqlnd as a requirment
I am using PHP7.4 on CentOS 8 and MySQL 8, and was getting an error about not being able to connect due to missing driver or unmet dependency. I wasn't able to find any errors in the logs that pointed me in the right direction, but found something on a site where someone mentioned that. As soon as I installed it I was able to connect.